#59 #60  Early fixes - improved coverage
Added unit tests for neutron service
Separated out all VLAN related stuff in a separate domain/service
moved refresh_ip out of connectivity.
